backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
psycopg2-binary>=2.9.1,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.0.6
email-validator>=1.1.3,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
python-barcode==0.15.1
Pillow==10.2.0
reportlab==4.0.9
redis==5.0.1
celery==5.3.6 